+++ public/index.html
@@ -100,14 +100,7 @@
     <div id="leaderboard-container" style="display: none;"></div>
 
     <!-- Include the transpiled scripts -->
-    <script defer src="dist/firebaseConfig.js"></script>
-    <script defer src="dist/util.js"></script>
-    <script defer src="dist/modeHandlers.js"></script>
-    <script defer src="dist/sessionHandlers.js"></script>
-    <script defer src="dist/questionHandlers.js"></script>
-    <script defer src="dist/resultsHandler.js"></script>
-    <script defer src="dist/initialization.js"></script>
-    <script defer src="dist/app.js"></script>
+    <script defer src="dist/bundle.js"></script>
 </body>
 
 </html>
