--- src/firebaseConfig.js
@@ -1,21 +1,0 @@
-import { getAnalytics } from "firebase/analytics";
-import { initializeApp } from "firebase/app";
-import { getFirestore } from "firebase/firestore";
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyDTFpB2roR1dV2xDVNDpig6ykuco5iNYCU",
-  authDomain: "calibration-trivia.firebaseapp.com",
-  projectId: "calibration-trivia",
-  storageBucket: "calibration-trivia.appspot.com",
-  messagingSenderId: "65056414205",
-  appId: "1:65056414205:web:e453dea073898b36f42194",
-  measurementId: "G-KLBZBZB995"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
-const db = getFirestore(app);
-
-export { analytics, db };
